--- app/src/main/java/woowacourse/shopping/data/ProductRepository.kt
@@ -1,7 +1,0 @@
-package woowacourse.shopping.data
-
-import woowacourse.shopping.model.Product
-
-interface ProductRepository {
-    fun getAllProducts(): List<Product>
-}
