+++ app/src/main/java/woowacourse/shopping/domain/model/Product.kt
@@ -1,3 +1,3 @@
-package woowacourse.shopping.model
+package woowacourse.shopping.domain.model
 
 class Product(val name: String, val price: Int, val imageUrl: String)
