+++ app/src/main/java/woowacourse/shopping/local/dao/CartProductDao.kt
@@ -1,8 +1,9 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.local.dao
 
 import androidx.room.Dao
 import androidx.room.Insert
 import androidx.room.Query
+import woowacourse.shopping.local.entity.CartProductEntity
 
 @Dao
 interface CartProductDao {
