+++ app/src/main/java/woowacourse/shopping/local/db/ShoppingDatabase.kt
@@ -1,7 +1,9 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.local.db
 
 import androidx.room.Database
 import androidx.room.RoomDatabase
+import woowacourse.shopping.local.dao.CartProductDao
+import woowacourse.shopping.local.entity.CartProductEntity
 
 
 @Database(entities = [CartProductEntity::class], version = 1, exportSchema = false)
