+++ app/src/main/java/woowacourse/shopping/ui/cart/CartViewModel.kt
@@ -3,8 +3,8 @@ package woowacourse.shopping.ui.cart
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import woowacourse.shopping.model.Product
-import woowacourse.shopping.data.CartRepository
+import woowacourse.shopping.domain.model.Product
+import woowacourse.shopping.domain.repository.CartRepository
 
 class CartViewModel(
     private val cartRepository: CartRepository,
