+++ app/src/main/java/woowacourse/shopping/ui/product/ProductViewHolder.kt
@@ -1,10 +1,10 @@
-package woowacourse.shopping.ui
+package woowacourse.shopping.ui.product
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.shopping.databinding.ItemProductBinding
-import woowacourse.shopping.model.Product
+import woowacourse.shopping.domain.model.Product
 
 class ProductViewHolder(
     private val binding: ItemProductBinding,
